Remove openapi stuff
Was making using middlewares harder and I think it was just overall
overkill. I think for what I want to do I can do manual tranlsation
of the zod types to something rust can work with or I can get fance and
convert zod to serde somehow or just handroll a smithy model and generate
the client that way.