Set up basic functionality and fix auth
Bearer auth was broken before. It would pass 100% of the time because
it was checking if a promise was false which it never would be.
Once that was fixed, had to add a new package for doign url encoding
and decoding casue my custom functions (copied from online) were busted
and not working. Once that was fixed, made some updates to do basic posting
and then reading a single post too.